requirements.txt
python-binance==1.0.17
ta==0.10.2
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
openai==1.30.1
groq==0.5.0
streamlit==1.35.0